完善接口功能，添加网络示例

+++ lib_acl_cpp/samples/cgi_upload/main.cpp
@@ -179,7 +179,8 @@ public:
 		// 读取 HTTP 客户端请求数据
 		while (len > 0)
 		{
-			k = len > sizeof(buf) ? sizeof(buf) : len;
+			k = len > (long long int) sizeof(buf)
+				? (long long int) sizeof(buf) : len;
 			ret = in.read(buf, k, false);
 			if (ret == -1)
 			{
